CL-Sort Sorting Package: * Move changelist_sort/sorting/developer_changelist.py to changelist_sort/sorting/sorting_changelist.py - replace tuple with list * Move changelist_sort/sorting/developer_file_pattern.py to changelist_sort/sorting/sorting_file_pattern.py - add dataclass annotation and migrate from kwargs to None fields * Update changelist_sort/sorting/developer_sort.py - rename class DeveloperChangelist to SortingChangelist, replace DeveloperChangelist tuples with SortingChangelist lists * Update changelist_sort/sorting/sort_mode.py - remove Developer SortMode
