+++ changelist_sort/sorting/sort_mode.py
@@ -5,5 +5,4 @@ from enum import Enum, auto
 
 class SortMode(Enum):
     MODULE = auto()
-    DEVELOPER = auto()
     SOURCESET = auto()
